pyproject.toml
[project]
name = "lsn_measurements"
version = "2025.04.17"
requires-python = ">=3.12"
dependencies = [
    "QtPy",
    "pathvalidate",
    "numpy",
    "scipy",
    "pyqtgraph",
    "pandas",

    # for quantities with units
    "astropy",

    # for communication with hardware
    "netifaces",
    "requests",
    "nidaqmx>=1.0.0",
    "pyserial",
    "PyVISA",
    "python-vxi11",
    "telnetlib3",
    "PyMeasure",
]
[project.optional-dependencies]
matplotlib = ['matplotlib']
formatting = [
    'black',
    'ruff<0.4.3; platform_system=="Windows" and platform_release<"10"',
    'ruff; platform_system!="Windows" or platform_release>="10"',
]
pyside6 = ['PySide6_Essentials >=6.1, !=6.8.0, !=6.8.0.1']
pyqt5 = ['PyQt5 >=5.15']
pyqt6 = ['PyQt6 >=6.1']

[tool.black]
line-length = 120

[tool.ruff]
line-length = 120

[tool.ruff.lint]
fixable = ["ALL"]
unfixable = []
extend-select = ["B", "D", "I", "ANN", "NPY", "RET", "SIM", "UP"]
extend-ignore = ["D1", "EM1", "N8", "PLR", "TRY003"]

[tool.ruff.lint.isort]
case-sensitive = true
combine-as-imports = true
